+++ config/dev.exs
@@ -1,2 +1,4 @@
 use Mix.Config
 config :logger, level: :info
+config :logger, handle_sasl_reports: true
+config :logger, handle_otp_reports: true