+++ config/test.exs
@@ -1,3 +1,3 @@
 use Mix.Config
 config :logger, level: :debug
-config :logger, handle_sasl_reports: true
+# config :logger, handle_sasl_reports: true